Change the type of FlowAnalysisData to simpilfy code
